1) Added feature to move (drug-and-drop) issue in the board - changed appropriated Controller actions, handlers and repo methods. 2) Fixed models and dtos. 3) Fixed queries and commands for handlers. 4) Added migration - 03-SprocsForBusinessLogic to DB.

+++ Board.BusinessLogic/DTOs/Issues/IssueResponseDto.cs
@@ -4,13 +4,13 @@ using Board.DataAccess.Models;
 namespace Board.BusinessLogic.DTOs.Issues;
 
 public record IssueResponseDto(
-    int Id,
+    long Id,
     string Title,
     string Description,
     DateTime? DueDate,
     DateTime CreatedAt,
-    int PositionInColumn,
-    int ColumnId,
+    long PositionInColumn,
+    long ColumnId,
     string CreatorId,
     string? AssigneeId);
